perception/requirements.txt
carla==0.9.15
gmpy2==2.2.1
imagecorruptions==1.1.2
imageio==2.35.1
matplotlib==3.7.5
mmcv==2.2.0
mmdet==3.3.0
mmdet3d==1.4.0
mmengine==0.10.3
mock==5.2.0
numba==0.48.0
numpy==1.23.4
object_detection==0.0.3
open3d==0.14.1
opencv_python==4.10.0.84
Pillow
pycocotools==2.0.7
PyYAML==6.0.1
scipy
seaborn==0.13.2
setuptools==68.2.2
Shapely==2.0.7
six==1.16.0
scikit-image
tqdm==4.66.5
ultralytics==8.3.18